+++ src/cetech/playground/private/property_editor.cpp
@@ -37,7 +37,7 @@ static ct_property_editor_a0 property_inspector_api = {
 };
 
 
-static void on_debugui(uint64_t bus_name,
+static void on_debugui(uint32_t bus_name,
                        void *event) {
     if (ct_debugui_a0.BeginDock(WINDOW_NAME,
                                 &_G.visible,
@@ -49,7 +49,7 @@ static void on_debugui(uint64_t bus_name,
     ct_debugui_a0.EndDock();
 }
 
-static void on_menu_window(uint64_t bus_name,
+static void on_menu_window(uint32_t bus_name,
                            void *event) {
     ct_debugui_a0.MenuItem2(WINDOW_NAME, NULL, &_G.visible, true);
 }
